feat(day 11): code skeleton for day 11
Added code skeleton mostly filled out for day 11. This implements a BFS in
the graph, keeping track of states we have already visited in a cache. I
just need to finish the current methods and fill out the remaining empty
methods to check. This is likely to be very inefficient, but will reserve
judgement until this is tested.