feat(day11): efficiencies
Pruning the possible states when we generate them speed sup part 1 by
double to 15s. Also refactored caching so that we store it in different
variables to try and improve the speed here. The current implementation
still doesn't work for part 2. Need to debug to see where the bottlenecks
are. If memory, we can replace strings in states with a mapping from int
-> str.