feat(day 11): add solution for part 2
This change removes the unnecessary complicated caching added in earlier,
and adds more intelligent tree pruning. We now cache on pair location
indexes, rather than named items and their positions in the system. By
ignoring names, and caching on indexes for pairs, we can get a lot more
cache hits (there may be states which is the same situation except one
element type is swapped with another) which means we can prune the tree
significantly.