feat(day 19): continueing changing solution for part 2
The old solution is far too slow, splicing the array each time takes too
long (up to n^2). The next idea is to go round the circle once, recording
every elf which is removed from the circle, but not removing them. This
is a bit more complicated as we have to be careful with indexes but should
be a lot quicker. The current solution does not currently work and
requires refactoring.