feat(day 22): print out the node grid
Added code to print out the node grid for part 2. Tiss prints out the
start, the goal, the empty node and any nodes whose capacity is too large
to be able to transfer data. From this, we can work out the part 2
solution by hand as this is a slider problem.