feat(day 18): part 2
Converted the part 1 solution to work for both part 1 inputs and part 2 inputs, but supporting multiple robots at difference entrances.

Had a really difficult time because my cache broken the solution - it turns out that the cache was writing "undefined" to the cache key so I had false hits. This is why I should be using flow or typescript!! (Or even better, writing unit tests)